fix: resolve implicit narrowing cast in compound assignments
- NodeView: use float literal 0.40f instead of double 0.40
- MapView: change mPaintingTime field from int to long
- RootMainView: explicit (int) cast for width multiplication
- Base64Coding: explicit (byte) cast in decode64 bit operations

+++ freemind/freemind/view/mindmapview/MapView.java
@@ -262,8 +262,8 @@ public class MapView extends JPanel implements ViewAbstraction, Printable, Autos
 	private boolean fitToPage = true;
 	private int pagesWide = 1;
 
-	int mPaintingTime;
-	int mPaintingAmount;
+	long mPaintingTime;
+	long mPaintingAmount;
 	static boolean printOnWhiteBackground;
 	static Color standardMapBackgroundColor;
 	static Color standardSelectColor;
